apps/backend/requirements.txt
flask==3.1.0
flask-cors==3.0.10
openai
python-dotenv
gunicorn